Add methods `translateState`, `translateCollector` and `translateIntegration` in I18n
